labs/lab_3/solution/lab3/pyproject.toml
[tool.poetry]
name = "lab3"
version = "0.3.0"
description = ""
authors = ["James Winegar <jameswine@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "^0.85"
uvicorn = "^0.18"
joblib = "^1.1"
scipy = "^1.9"
scikit-learn = "1.1.2"
requests = "^2.28"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}

[tool.poetry.dev-dependencies]
pytest = "^7.0"
isort = "^5.10.1"
black = "^22.1"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
